# docs/collection.swagger.yaml
openapi: 3.0.0
info:
  title: LOCAL MUSIC MANAGER - API
  version: 1.0.0
servers:
  - url: http://127.0.0.1:5000
components:
  securitySchemes:
    noauthAuth:
      type: http
      scheme: noauth
    bearerAuth:
      type: http
      scheme: bearer
...
paths:
...
  /api/v1/songs/{songId}:
    get:
      parameters:
        - name: songId
          in: path
          schema:
            type: string
          required: true
          example: HKNqyiV38ss
      tags:
        - Songs
      summary: Get song
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...